ignore already-set cert if it's in nix store
Trying to reinstall daemon over no-daemon installs causes a
little trouble for this code that supports user overrides of
NIX_SSL_CERT_FILE. The code would see the single-user env as
an override, and skip installing a cert into the default
profile.

I saw two fixes to this:
1) don't do anything special if the env is set, but it points
   at a file in the Nix store (this is implemented here)
2) go ahead and keep the user's override, but also always go
   ahead and install the default-profile's cert (normal
   process, minux exporting a new NIX_SSL_CERT_FILE).

I'm not sure which is better, but I went to all of the trouble
of figuring out how to canonicalize the path with stock macos
tools--so I figured I should commit the hard one first before
I forget.